appended patch stamp
Put an insert statement like this at the end of each patch script. This tracks which patches have been applied. The timestamp marks what version of a patch was applied, as long as there was a git pull immediately before running the .sql. Modify the comment if anything different than the patch script. The insert goes at the end, not the top, of the .sql file so if it fails it will not get "stamped". We could go so far as a transaction if partial execution gets messy.