minor edits to README.md
Needed to make it clear when the front page said "last updated" that this referred to the README, not to the repo as a whole. Also added encouragement for users to keep in contact, to ease their experience and to emphasize the development is not yet as stable as we would like prior to sites using mb for production. I did not mention the eml2mb repo soon to be created but we may want to add that soon, once it is up and ready for visitors.