add 27_abstract_text_type
Two places in eml the complex text type is handled by metabase as a choice of plain text, markdown (only with eml2.2) or a filename of a Word doc external to metabase. (There is also a special case for content as xml, not described here.) Following the code pattern used with MethodStep's Description and DescriptionType column, I added an AbstractType column to the DataSet table (which already had the Abstract column). 

I added a column abstract_type to the corresponding view. I noticed one column in that view had an outgoing column name inconsistent with our pattern of using eml element names all lowercase, so renamed UpdateFrequency as maintenanceupdatefrequency. 

(Special note to @atn38 do you want the AbstractType column to be in order in DataSet to the left of Abstract? The ALTER will put the new column at the end but you can make magic happen with OBF, if worth the effort.)